feat: database created (ongoing)
