+++ zenhabits_app/pubspec.yaml
@@ -35,9 +35,10 @@ dependencies:
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.8
   # Floor 
-  floor: ^2.0.0
-  floor_annotation: ^2.0.0
-  sqflite: ^2.0.0
+  floor: ^1.4.2
+  sqflite: ^2.0.0+4
+  path_provider: ^2.0.11
+  path: ^1.8.2
 
 dev_dependencies:
   flutter_test:
@@ -50,8 +51,8 @@ dev_dependencies:
   # rules and activating additional ones.
   flutter_lints: ^5.0.0
 
-  build_runner: ^2.0.0
-  floor_generator: ^2.0.0
+  floor_generator: ^1.4.2
+  build_runner: ^2.1.2
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
